se hicieron cambios en el metodo verificarPosibilidadDeCompra y se creo el servicio para enviar el correo

+++ backend/ladrillera/src/main/java/com/ladrillera/ladrillera/ventas/repository/VentasRepository.java
@@ -16,23 +16,26 @@ import java.util.Map;
 public interface VentasRepository extends JpaRepository<Ventas, Long> {
 
         // Método para buscar ventas por clienteId
-        List<Ventas> findByClienteId(Integer clienteId);
+        List<Ventas> findByClienteId(Long clienteId);
+
 
         // Contar ventas por cliente en un día
         @Query("SELECT COUNT(v) FROM Ventas v WHERE v.clienteId = :clienteId AND v.fecha = :fecha")
-        long countByClienteIdAndFecha(@Param("clienteId") Integer clienteId, @Param("fecha") LocalDate fecha);
+        long countByClienteIdAndFecha(@Param("clienteId") Long clienteId, @Param("fecha") LocalDate fecha);
 
         // Contar ventas por cliente entre un rango de fechas (mes)
         @Query("SELECT COUNT(v) FROM Ventas v WHERE v.clienteId = :clienteId AND v.fecha BETWEEN :startDate AND :endDate")
-        long countByClienteIdAndFechaBetween(@Param("clienteId") Integer clienteId,
-                        @Param("startDate") LocalDate startDate,
-                        @Param("endDate") LocalDate endDate);
+        long countByClienteIdAndFechaBetween(@Param("clienteId") Long clienteId,
+                                             @Param("startDate") LocalDate startDate,
+                                             @Param("endDate") LocalDate endDate);
 
         // Contar ventas por cliente en un rango de fechas (año)
         @Query("SELECT COUNT(v) FROM Ventas v WHERE v.clienteId = :clienteId AND v.fecha BETWEEN :startDate AND :endDate")
-        long countByClienteIdAndFechaAnual(@Param("clienteId") Integer clienteId,
-                        @Param("startDate") LocalDate startDate,
-                        @Param("endDate") LocalDate endDate);
+        long countByClienteIdAndFechaAnual(@Param("clienteId") Long clienteId,
+                                           @Param("startDate") LocalDate startDate,
+                                           @Param("endDate") LocalDate endDate);
+
+
 
         // Método para contar ventas por sucursal anualmente
         @Query("SELECT COUNT(v) FROM Ventas v WHERE v.sede = :sede AND v.fecha BETWEEN :startDate AND :endDate")
